+++ Bookshop/src/main/java/com/bookshop/controller/ProductController.java
@@ -4,7 +4,6 @@ import com.bookshop.model.Product;
 import com.bookshop.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
 @RestController
@@ -14,15 +13,16 @@ public class ProductController {
 	@Autowired
 	public ProductService productService;
 
-	@GetMapping
+		@GetMapping
 	public List<Product> getListOfProduct() {
 		return productService.getAllProducts();
 	}
 
 	@GetMapping("/{productId}")
-	public Product getProductById(@PathVariable Integer productId) {
+	public Product getProductById(@PathVariable Long productId) {
 		return productService.getProductById(productId);
 	}
+
 
 	@PostMapping
 	public Product addProduct(@RequestBody Product product) {
@@ -30,12 +30,12 @@ public class ProductController {
 	}
 
 	@PutMapping("/{productId}")
-	public Product editProduct(@RequestBody Product product, @PathVariable Integer productId) {
+	public Product editProduct(@RequestBody Product product, @PathVariable Long productId) {
 		return productService.editProduct(product, productId);
 	}
 
 	@DeleteMapping("/{productId}")
-	void deleteProductById(@PathVariable Integer productId) {
+	void deleteProductById(@PathVariable Long productId) {
 		productService.deleteProductById(productId);
 	}
