+++ Bookshop/src/main/java/com/bookshop/repository/StoreItemRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoreItemRepository extends JpaRepository<StoreItem, Integer> {}
+public interface StoreItemRepository extends JpaRepository<StoreItem, Long> {}
